KD: changed XpsiA, XpsiB, XpsiD arrays in multi.h to double
git-svn-id: svn+ssh://savannah01.psi.ch/repos/gems/GEMSELEKTOR/trunk@158 7d9feeff-4400-0410-95ed-9cb8b92a518d

+++ gems2/mods/submod/ipm_gamma.cpp
@@ -1087,8 +1087,8 @@ double TProfil::Ej_init_calc( double, int j, int k)
         /* get ist - index of surface type and isp - index of surface plane  */
         ist = pmp->SATNdx[j][0] / MSPN;
         isp = pmp->SATNdx[j][0] % MSPN;
-        psiA = double(pmp->XpsiA[k][ist]);
-        psiB = double(pmp->XpsiB[k][ist]);
+        psiA = pmp->XpsiA[k][ist];
+        psiB = pmp->XpsiB[k][ist];
         Ez = double(pmp->EZ[j]);
         if( !isp )  /* This is  A (0) plane */
             F0 += psiA * Ez * pmp->FRT;
